MainActivity: relaunch bigme Launcher onStart
Bigme launcher controls the edge shortcuts, on Bigme devices.
These shortcuts in turn allows to move across applications.
When locked and unlocked the Launcher gets killed; for this reason
the shortcuts stop to work. By restarting the process `onStart` we
ensure that the original Launcher is active in background and normal
shortcuts works properly.